third_party/blink/web_tests/editing/input/text-input-controller-leak-document-expected.txt

PASS markedRangeResultGC.wasCollected is true
PASS selectedRangeResultGC.wasCollected is true
PASS firstRectForCharacterRangeGC.wasCollected is true
PASS successfullyParsed is true

TEST COMPLETE
